services/api/pyproject.toml
[project]
name = "workout-tracker-api"
version = "0.1.0"
description = "Workout Tracker API Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pyjwt>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]
